+++ src/utils/spid.ts
@@ -1,4 +1,12 @@
-import { fromEither, fromNullable, Option, tryCatch } from "fp-ts/lib/Option";
+import { QueueClient } from "@azure/storage-queue";
+import { format as dateFnsFormat } from "date-fns";
+import {
+  fromEither,
+  fromNullable,
+  isNone,
+  Option,
+  tryCatch
+} from "fp-ts/lib/Option";
 import * as t from "io-ts";
 import { UTCISODateFromString } from "italia-ts-commons/lib/dates";
 import {
@@ -6,6 +14,7 @@ import {
   NonEmptyString,
   PatternString
 } from "italia-ts-commons/lib/strings";
+import { log } from "./logger";
 
 const SAML_NAMESPACE = {
   ASSERTION: "urn:oasis:names:tc:SAML:2.0:assertion",
@@ -64,3 +73,65 @@ export const SpidBaseMsg = t.interface({
 });
 
 export type SpidBaseMsg = t.TypeOf<typeof SpidBaseMsg>;
+
+export const makeSpidLogCallback = (queueClient: QueueClient) => (
+  sourceIp: string | null,
+  payload: string,
+  payloadType: "REQUEST" | "RESPONSE"
+): void => {
+  const maybeXmlPayload = tryCatch(() =>
+    new DOMParser().parseFromString(payload, "text/xml")
+  );
+  if (isNone(maybeXmlPayload)) {
+    log.error(`SpidLogCallback|ERROR=Cannot parse SPID XML Payload`);
+    return;
+  }
+
+  const xmlPayload = maybeXmlPayload.value;
+
+  const maybeRequestId =
+    payloadType === "REQUEST"
+      ? getRequestIDFromRequest(xmlPayload)
+      : getRequestIDFromResponse(xmlPayload);
+
+  const maybeFiscalCode = getFiscalNumberFromPayload(xmlPayload);
+
+  if (isNone(maybeRequestId)) {
+    log.error(`SpidLogCallback|ERROR=Cannot get Request ID from XML Payload`);
+    return;
+  }
+
+  if (isNone(maybeFiscalCode) && payloadType === "RESPONSE") {
+    log.error(
+      `SpidLogCallback|ERROR=Cannot recognize fiscal Code on XML Payload provided by SAMLResponse`
+    );
+    return;
+  }
+
+  const item = {
+    fiscalCode: maybeFiscalCode.toUndefined(),
+    spidRequestId: maybeRequestId.toUndefined()
+  };
+
+  const baseMsg: SpidBaseMsg = {
+    createdAt: new Date(),
+    createdAtDay: dateFnsFormat(new Date(), "YYYY-MM-DD"),
+    ip: sourceIp as IPString,
+    payload,
+    payloadType
+  };
+
+  const spidMessage = {
+    ...baseMsg,
+    ...item
+  };
+
+  const spidMsg = Buffer.from(JSON.stringify(spidMessage)).toString("base64");
+
+  // we don't return the promise here
+  // the call follows fire & forget pattern
+  queueClient.sendMessage(spidMsg).catch(err => {
+    log.error(`SpidLogCallback|ERROR=Cannot enqueue SPID payload`);
+    log.debug(`SpidLogCallback|ERROR_DETAILS=${err}`);
+  });
+};
